Refactor: Move QuerySecrets method to k8s_utils module
